Bluetooth: Mesh: Model: Fix last temp storage
Fixes bug in temprature server where the last stored temprature value is
incorrectly set to the start value for a level delta set message. To
make it possible to do corrective deltas in this call, the start value
has previously been stored in the last value state, assuming that the
actual last value has already been saved to persistent storage.
When the CONFIG_BT_MESH_MODEL_SRV_STORE_TIMEOUT is set to a
value greater than 0 this assumption is not valid.

Signed-off-by: Anders Storrø <anders.storro@nordicsemi.no>