lib: fprotect: Delete Kconfig option FPROTECT_ENABLE_IS_PROTECTED
I can't see any reason for the Kconfig option
FPROTECT_ENABLE_IS_PROTECTED to exist so it is removed.

It is best-practice to rely on the linker to exclude unused functions
from the binary instead of using Kconfig options for every single
function.

This is a breaking change for anyone that is configuring
FPROTECT_ENABLE_IS_PROTECTED explicitly, but I don't expect anyone is
actually doing this.

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>