Bluetooth: Mesh: Invalidate scene when state changed by state machine
From MshMDLv1.0.1, section 5.1.3.2.1:

When any of the element's state that is marked as “Stored with Scene”
has changed not as a result of a Scene Recall operation, the value of
the Current Scene state shall be set to 0x0000.

Thus, when Light LC State Machine changes Light OnOff state, the scene
shall be invalidated.

Signed-off-by: Pavel Vasilyev <pavel.vasilyev@nordicsemi.no>