net: lib: azure_iot_hub: Use correct type when receiving the device twin
After a connection has been established, the azure iot hub client
automatically requests the entire device twin document.

When the device twin response is received it is marked with the
`AZURE_IOT_HUB_TOPIC_TWIN_DESIRED` topic type which is wrong.
This type indicates that the response contains only the `desired`
section of the device twin.

Use the `AZURE_IOT_HUB_TOPIC_TWIN_REQUEST_RESULT` topic type
to correctly identify the payload, and to make sure that the correct
event `AZURE_IOT_HUB_EVT_TWIN_RECEIVED` is sent to the application.

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>