applications: asset_tracker_v2: Handle lwm2m FOTA state changes
Handle lwm2m FOTA state changes and propagate the correct FOTA event to
the rest of the application. This allows the application to display
the correct FOTA LED patterns, and to detect when the FOTA has been
performed and a reboot is required.

Add Kconfig option `CONFIG_LWM2M_CLIENT_UTILS_FIRMWARE_UPDATE_REBOOT`
that gives the application the opportunity to disable
rebooting from the firmare object when it wants to handle
the reboot it self.

Update LWM2M integration test.

Fixes CIA-644

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>