samples: wifi: shell: Add overlays for nRF52840 and nRF9160 DKs
Add overlays that allow running the sample on nRF52840 and nRF9160
DKs with the nRF7002 EK shield. Some peripherals enabled by default
on those DKs generate conflicts on lines that are used by nRF7002 EK,
so these peripherals are disabled. For nRF9160 DK, additionally SPI
pins are configured with high drive so that SCK frequency > 4 MHz can
be used.

Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>