samples: bluetooth: Set report ID in HIDs keyboard to 0.
It is confirmed with iOS 16 that setting the report ID for
input and output reports to 0 in peripheral_hids_keyboard
sample fixes an issue where the device would not be detected
as a native keyboard input.
The new setting seems to be aligned also with the
HIS Service Specification V10r00, section 2.5.3.2
which says the ID shall be non-zero, when there are multiple
reports defined within a specific report type. In this case
there is just 1 input and 1 output report, so there is no need
to distinguish reports within each type by their IDs.

Ref. NCSIDB-831

Signed-off-by: Michał Grochala <michal.grochala@nordicsemi.no>