net: lib: lwm2m_client_utils: Check credential length before writing
On some corner cases, the code could try to convert non-existent
PSK key to a proper format and write that to modem. Added a
length check.

Also, sometimes, if modem did have keys, but we had not provisioned
it, the code tried to write non-existing key to a modem. Fixed.

Signed-off-by: Seppo Takalo <seppo.takalo@nordicsemi.no>