applications: asset_tracker_v2: tests: Fix pointer expects in unit test
- Skip checking the mode parameter of lte_lc_func_mode_get() as
this is a pointer to a local variable. Instead return the memory
location of a test local variable through the pointer.
- Skip checking the value parameter of lwm2m_engine_get_u8() as this is
also a pointer to a local variable. No need to return anything through
pointer in this case as this is out of test scope.

This fixes test failures which happen when using a slightly different
compiler version or when building with no optimizations.

Signed-off-by: Balaji Srinivasan <balaji.srinivasan@nordicsemi.no>