security: Basic PSA certification level 2 support
This adds the basic features required for PSA level 2 certification.

bl_storage:
Support storing the life cycle state in OTP memory of the
UICR is done in header implementation, so TF-M as non-zephyr image can
also, use the same code by just including the header.
Storing an implementation in OTP
Adding support to include the library in other images, making it
possible to run the provisioning sample without a b0 bootloader, as the
b0 bootloader contains secrets that must not be on the device during
provisioning state.

TF-M sample:
Also, this adds a TF-M provisioning sample, that generates and stores
the hardware unique keys (including MKEK) and the Initial attestation
key to transitioning the device to the provisioning life cycle state.

TF-M module:
Add support to read and change the life cycle state from TF-M.
Add support for checking if the device was provisioned and either halt
booting of the non-secure application or transition the device into
the secured life cycle state.

Ref: NCSDK-17500
Ref: NCSDK-17001

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>
Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>
Signed-off-by: Markus Swarowsky <markus.swarowsky@nordicsemi.no>