tfm: Set the TF-M non-secure application as the app
Set the TF-M non-secure application as the main app.hex file instead
of defining a hex file for the tfm_nonsecure partition.
While these are the same partitions replacing the application hex
file allows the TF-M non-secure application to be properly signed and
flashed when CONFIG_BOOTLOADER_MCUBOOT is enabled.

NCSDK-17680

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>