samples: wifi: shell: Add OTBR support
Add an overlay for Wi-Fi thread combo, useful for the OTBR [1].

OpenThread interface is the first interface though its not default, so,
we need to explicitly initialize the network stack to use Wi-Fi
interface else DHCPv4 chooses OpenThread and it will drop the packets as
it only supports IPv6.

Also, as we are explicitly initializing the network stack, disable
automatic initialization, else it causes boot hang.

[1] - https://openthread.io/guides/border-router

Signed-off-by: Krishna T <krishna.t@nordicsemi.no>