sysbuild: cmake: partition manager integration
This commit ports and updates cmake/partition_manager.cmake to a
sysbuild specific version in cmake/sysbuild/partition_manager.cmake.

The main reason for a new CMake file is the need to run partition
manager outside of an application image build.
The parent-child image scheme runs partition manager in the parent image
build of each core.

Sysbuild builds each image individually and the concept of a parent
image of a core does not exist.
The sysbuild partition manager has been updated to provide functions
which allows partition manager to be invoked multiple times in same
scope, once per domain (core).

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>