samples: nrf5340: multiprotocol_rpmsg: decrease BT_MAX_CONN
The multiprotocol_rpmsg is hovering around the limits of RAM available
on nRF5340 network core. In order to free some of the memory and make it
easier to extend the sample with additional features, this commit
reduces the maximum number of concurrent BLE connections to 4. In
real-life scenarios, it is unexpected for more concurrent connections to
be needed*. The decrease should therefore impose no actual limitations
on the users while saving a lot of RAM.

*in fact, is is uncommon for any multiprotocol application to use more
concurrent connections than 2, so there's still a margin.

Signed-off-by: Jędrzej Ciupis <jedrzej.ciupis@nordicsemi.no>