net: lib: lwm2m_client_utils: Connmon Update
Update the connectivity monitor whenever we get the
cell update event from the link control and the
device is connected to the network. The event
itself only contains cell id and tracking area code,
so we need to rely on modem info for getting all the
data. RSRP is handled separately as it has its own
callback.

Signed-off-by: Jarno Lämsä <jarno.lamsa@nordicsemi.no>