doc: Drop read only functionality for bl_storage library from app
If the application rely on reading out the OTP based on the
bl_storage format, the NISB and the application must always be compiled
with the same version of the bl_storage library.
But as the format of OTP within the bl_storage might change in future,
it is likely that the application is build with a newer version causing
read failures if a DFU is performed.

Therefore we remove the feature to use the library within the
application.

Signed-off-by: Markus Swarowsky <markus.swarowsky@nordicsemi.no>