samples: matter: apply fix for increased flash wear
MoveToLevel command of LevelControl cluster, used in light
bulb sample, causes a fast change of CurrentLevel attribute
that is persisted to flash. It leads to lots of flash writes
that needlessly utilizes the flash.

Use the deferred attribute persister for CurrentLevel
attribute to write it to flash only when it remains constant
for 5 seconds.

Signed-off-by: Damian Krolik <damian.krolik@nordicsemi.no>