net: fota_download: Save 6K flash by not using fd for state checking
Initializing the fd member of the download_client struct resulted in a
6K increase in flash size of all applications using this module. The
purpose of initing the fd variable to -1 was to make the
fota_download_cancel return error in case the download was not in
prgoress. It is now made to re-use the existing downloading variable to
do the same.

The reason for 6K increase in flash was that intializing the member
(in this case fd) to -1 involves the linker placing the
entire struct in both RAM and flash (the init value is stored in flash
and the CPU will populate the struct in RAM with the init value in
flash during startup).

Also modified the tests to ensure that the API fota_download_cancel
returns -EAGAIN when a download is not in progress.
Added a new test to ensure that fota_download_cancel() returns error
when it was called without init.

As a side effect, this uncovered two bugs which are also fixed in this.
1) download_client_callback was continuing to read offset value even if
dfu_target_offset_get returned fail. This can lead to undefined
behavior.
2) The state "downloading" was not set to false when an error event
was raised. This lead to a dependency on the app calling
fota_download_cancel() to cleanup the state.

Signed-off-by: Balaji Srinivasan <balaji.srinivasan@nordicsemi.no>