kconfig: optional sourcing of BOARD_DIR in nrf
Add Kconfig osource of BOARD_DIR prefixed with ZEPHYR_NRF_MODULE_DIR.

A regular build will set BOARD_DIR to an absolute path, and thus the
prefixing with ZEPHYR_NRF_MODULE_DIR will point to an invalid path.
This is expected and desired because the actual BOARD_DIR is sourced
from Zephyr repository Kconfig tree and we don't want to source it
twice.

For documentation build and compliance, the BOARD_DIR is a glob pattern
and this commit ensures that board Kconfig files in the sdk-nrf tree are
sourced in those cases.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>