cmake: update PROTOC setting to cache type STRING
Fixes: NCSIDB-816

Previous fix, e2662c11abcad726692d56df03a1730f267dec00, of NCSIDB-816
was incomplete as it didn't work properly on Windows OS.

Part of the reason is that `;` is separating elements in environment
setting PATH, but in CMake it acts as a element separator for lists.
This can normally be handled with proper escaping, but the variables
expanded inside other variables combined with the cache type FILEPATH
results in escaped `;` to be regular `;`.

This means that on Windows the PATH setting would always become a space
separated string upon custom command invocation.

This commit fixes above issues by:
- Escape `;` in the custom command path, this ensure that on Windows OS
  where `;` is used, then CMake will refrain from treating the string as
  a list.
- Use cache type STRING instead of FILEPATH
  This ensures CMake will refrain from manipulating the setting.
  Apparently the CMake processing of a FILEPATH setting to handle
  relative path also removes escaped `;` in the process.
- Use `cmake_path(CONVERT ...)`
  This allows to construct regular CMake lists of paths, and thereafter
  ensure the list is a proper OS separated list.
- Removed the use of PROTOC to set PROTOBUF_PROTOC_EXECUTABLE, as doing
  so removed one level of escaped `;`.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>