doc: update references to add_partition_manager_config
The function add_partition_manager_config was prefixed with nvs_ to get
exposed to OOT users to let them add partition manager configuration
files since the commit bb1da8c7e25e368341634114a2746f733bb763e3.

The references in the documentation were not updated since then, and it
still uses the former name.

This updates all the references for the function in the repository.

Jira: NCSDK-5493
Signed-off-by: Gaël PORTAY <gael.portay@rtone.fr>