tfm: Update TF-M configuration to generate a config header file
TF-M 1.7.0 has partial Kconfig support to generate a header file for
some of configurations
This commit skips the TF-M kconfig files and instead generates the
header file from our existing Kconfig entries in both sdk-zephyr and
sdk-nrf.
This is just a temporary solution until we have proper integration
with the TF-M Kconfig files.

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>