tfm: Align TF-M partitions with SPU regions when used with MCUboot
The tfm_nonsecure and sram_nonsecure partition must be SPU region
aligned, therefore thesize of the tfm partition must be reduced by
the size of the mcuboot_pad partition when TF-M is used in combination
with MCUBoot.

To ensure that is still the case if the partition file is change a
build assert is added to ensure that is still vaild.

Ref: NCSDK-19515

Signed-off-by: Markus Swarowsky <markus.swarowsky@nordicsemi.no>