scripts: west: s/shortlog/title/
I have incorrectly been referring to git commit titles as shortlogs.
Though git commit titles do appear in the 'git shortlog' output, a git
commit title is not a shortlog.

This commit affects nrfconnect/action-oss-history, and can only be
used with a version of that repository that can handle this change in
the JSON output.

Generated automatically from the west_commands directory with:

  $ rg -l shortlog | xargs sed -i s/shortlog/title/g

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>